remove lib/data_sources cycle
