+++ front/components/data_source/RequestDataSourceModal.tsx
@@ -11,9 +11,8 @@ import { useContext, useEffect, useState } from "react";
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
 import {
   getConnectorProviderLogoWithFallback,
-  getDataSourceName,
 } from "@app/lib/connector_providers";
-import { isManaged } from "@app/lib/data_sources";
+import { getDataSourceName, isManaged } from "@app/lib/data_sources";
 import { sendRequestDataSourceEmail } from "@app/lib/email";
 import logger from "@app/logger/logger";
